1. Additional code on all entities for update and delete of records, and return robust responses, like JsonResponse or HTTP exception, that can be handled by clients neatly. 2. a) All tests modified to handle according to the changed API endpoint Json and request bodies. With 3 types of clients, authorized_admin, authorized_client and normal client.     b) test_projects and test_audio were empty stubs, now with actual unit tests that exercise all CRUD operations. test_audio also exercises mp3 file uploads. 3.  app/utils.py  - "sa" for Sanskrit added to language enum.      app/projects.py -      a) get_project_by_id or by_name with 404 if not found b) JSONResponse with status code 201      app/isha/routers/sutras.py, meanings.py, transliterations.py, interpretations.py, bhashyams.py  - a) get_project_by_id or by_name with 404 if not found b) JSONResponse with status code 201 c) update_<entity> and delete_<entity> functions and return robust responses, like JsonResponse or HTTP exception, that can be handled by clients neatly.      app/kena/routers/ a) interpretations.py, bhashyams.py - Philosophy enum (adv, dva, vis) added   b) audio.py - filepath construction creates <chapter>_<sutranmber>_<mode>.<filetype> 4. added httpx and pytest to requirements.txt

+++ app/isha/routers/bhashyams.py
@@ -1,4 +1,5 @@
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, status, HTTPException
+from fastapi.responses import JSONResponse
 from sqlalchemy.orm import Session
 
 from app import models as app_models
@@ -6,7 +7,7 @@ from app import oauth2
 from app.database import get_db
 from app.errors import conflict_error_response, not_found_error_response
 from app.isha import models, schemas
-from app.utils import Language
+from app.utils import Language, Philosophy
 
 from .utils import get_sutra_or_404
 
@@ -14,30 +15,15 @@ router = APIRouter(prefix="/sutras", tags=["Bhashyams"])
 
 
 def get_bhashyam_or_404(sutra_id: int, language: Language, db: Session):
-    db_bhashyam = (
-        db.query(models.Bhashyam)
-        .filter(models.Bhashyam.sutra_id == sutra_id)
-        .filter(models.Bhashyam.language == language)
-        .first()
-    )
-
-    if not db_bhashyam:
-        not_found_error_response()
-
+    db_bhashyam = (db.query(models.Bhashyam).filter(models.Bhashyam.sutra_id == sutra_id).filter(models.Bhashyam.language == language).first())
+    if not db_bhashyam:not_found_error_response()
     return db_bhashyam
 
-
 @router.get("/{sutra_project}/{sutra_chapter}/{sutra_no}/bhashyam", response_model=schemas.BhashyamOut)
-def get_bhashyam(
-    sutra_project: str, sutra_chapter: int, sutra_no: int, lang: Language = Language.en, db: Session = Depends(get_db)
-):
+def get_bhashyam(sutra_project: str, sutra_chapter: int, sutra_no: int, lang: Language = Language.en, phil: Philosophy = Philosophy.advaita, db: Session = Depends(get_db)):
     sutra = get_sutra_or_404(sutra_project, sutra_chapter, sutra_no, db)
-
     bhashyam = get_bhashyam_or_404(sutra.id, lang, db)
-
     return bhashyam
-
-
 @router.post("/{sutra_project}/{sutra_chapter}/{sutra_no}/bhashyam", status_code=status.HTTP_201_CREATED)
 def add_bhashyam(
     sutra_project: str,
@@ -49,28 +35,21 @@ def add_bhashyam(
 ):
     # Retrieve sutra or raise 404 if not found
     sutra = get_sutra_or_404(sutra_project, sutra_chapter, sutra_no, db)
-
     # Check if bhashyam for the given sutra and language already exists
     db_bhashyam = (
         db.query(models.Bhashyam)
         .filter(models.Bhashyam.sutra_id == sutra.id)
         .filter(models.Bhashyam.language == bhashyam.language)
+        .filter(models.Bhashyam.philosophy == bhashyam.philosophy)
         .first()
     )
 
-    if db_bhashyam:
-        conflict_error_response(
-            f"Bhashyam for {sutra_project} sutra chapter {sutra_chapter} number {sutra_no} in language {bhashyam.language} already exists!"
-        )
-
+    if db_bhashyam:conflict_error_response(f"Bhashyam for {sutra_project} sutra chapter {sutra_chapter} number {sutra_no} in language {bhashyam.language} philosophy {bhashyam.philosophy} already exists!")
     new_bhashyam = models.Bhashyam(**bhashyam.model_dump(), sutra_id=sutra.id)
-
     db.add(new_bhashyam)
     db.commit()
     db.refresh(new_bhashyam)
-
     return {"id": new_bhashyam.id}
-
 
 @router.put("/{sutra_project}/{sutra_chapter}/{sutra_no}/bhashyam", status_code=status.HTTP_204_NO_CONTENT)
 def update_bhashyam(
@@ -79,6 +58,7 @@ def update_bhashyam(
     sutra_chapter: int=0,
     sutra_no: int=0,
     lang: Language='en',
+    phil: Philosophy="adv",
     db: Session = Depends(get_db),
     current_user: app_models.User = Depends(oauth2.get_current_user),
 ):
@@ -86,12 +66,13 @@ def update_bhashyam(
 
     db_bhashyam = get_bhashyam_or_404(sutra.id, lang, db)
 
-    # Update the fields
-    for key, value in bhashyam.model_dump().items():
-        setattr(db_bhashyam, key, value)
-
-    db.commit()
-    db.refresh(db_bhashyam)
+    if db_bhashyam:
+        # Update the fields
+        for key, value in bhashyam.model_dump().items(): setattr(db_bhashyam, key, value)
+        db.commit()
+        db.refresh(db_bhashyam)
+        return JSONResponse(status_code=status.HTTP_202_ACCEPTED, content=f'Updated sutra {sutra_project} chapter {sutra_chapter} sutra {sutra_no} text "{bhashyam.model_dump()["text"]}"')
+    else: raise HTTPException(status_code=404, detail=f"Project {sutra_project} chapter {sutra_chapter} sutra {sutra_no} bhashyam {bhashyam.model_dump()["language"]} not found")
 
 
 @router.delete("/{sutra_project}/{sutra_chapter}/{sutra_no}/bhashyam", status_code=status.HTTP_204_NO_CONTENT)
@@ -100,11 +81,15 @@ def delete_bhashyam(
     sutra_chapter: int,
     sutra_no: int,
     lang: Language,
+    phil: Philosophy,
     db: Session = Depends(get_db),
     current_admin: app_models.User = Depends(oauth2.get_current_admin),
 ):
     sutra = get_sutra_or_404(sutra_project, sutra_chapter, sutra_no, db)
     bhashyam = get_bhashyam_or_404(sutra.id, lang, db)
+    if bhashyam:
+        db.delete(bhashyam)
+        db.commit()
+        return JSONResponse(status_code=status.HTTP_200_OK, content=f"Deleted sutra {sutra_project} chapter {sutra_chapter} sutra {sutra_no} bhashyam")
+    else: raise HTTPException(status_code=404, detail=f"Project {sutra_project} chapter {sutra_chapter} sutra {sutra_no} bhashyam not found")
 
-    db.delete(bhashyam)
-    db.commit()
