+++ app/isha/routers/sutras.py
@@ -14,6 +14,14 @@ from .utils import get_sutra_or_404
 
 router = APIRouter(prefix="/sutras", tags=["Sutras"])
 
+def get_sutra_by_id_or_404(sutra_id: int, db: Session):
+    sutra = (db.query(models.Sutra).filter(models.Sutra.id == sutra_id).first())
+    if not sutra:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Sutra ID {sutra_id} not found",
+        )
+    return sutra
 
 @router.get("/", response_model=List[schemas.SutraListOut])
 def get_sutras(project_name: str = "isha", db: Session = Depends(get_db)):
@@ -28,8 +36,12 @@ def get_sutras(project_name: str = "isha", db: Session = Depends(get_db)):
 @router.get("/{sutra_project}/{sutra_chapter}/{sutra_no}", response_model=schemas.SutraOut)
 def get_sutra(sutra_project: str='isha', sutra_chapter: int=0, sutra_no: int=0, db: Session = Depends(get_db)):
     return get_sutra_or_404(sutra_project, sutra_chapter, sutra_no, db)
-
+@router.get("_by_id/{sutra_id}")
+def get_project(sutra_id: int, db: Session = Depends(get_db)):
+    sutra = get_sutra_by_id_or_404(sutra_id, db)
+    return sutra
 @router.post("/", status_code=status.HTTP_201_CREATED)
+
 def add_sutra(
     project: app_schemas.ProjectCreate,
     sutra: schemas.SutraCreate,
@@ -62,24 +74,26 @@ def add_sutra(
         content={"id": sutra.id},
     )
 
-
-@router.put("/{sutra_project}/{sutra_chapter}/{sutra_no}/{sutra_text}", status_code=status.HTTP_204_NO_CONTENT)
+@router.put("/{sutra_project}/{sutra_chapter}/{sutra_no}", status_code=status.HTTP_202_ACCEPTED)
+# @router.put("/", status_code=status.HTTP_204_NO_CONTENT)
 def update_sutra(
     sutra_update: schemas.SutraUpdate,
     sutra_project: str='isha',
     sutra_chapter: int=0,
     sutra_no: int=0,
-    sutra_text: str="?",
+    # sutra_text: str="?",
     db: Session = Depends(get_db),
     current_user: app_models.User = Depends(oauth2.get_current_user),
 ):
     sutra = get_sutra_or_404(sutra_project, sutra_chapter, sutra_no, db)
-
-    for key, value in sutra_update.model_dump().items():    
-        setattr(sutra, key, value)
-
-    db.commit()
-
+    if sutra:
+        sutra_update_items = sutra_update.model_dump()
+        sutra_text = sutra_update_items["text"]
+        for key, value in sutra_update_items.items():
+            setattr(sutra, key, value)
+        db.commit()
+        return JSONResponse(status_code=status.HTTP_202_ACCEPTED, content=f"Updated sutra {sutra_project} chapter {sutra_chapter} sutra {sutra_no} text {sutra_text}")
+    else: raise HTTPException(status_code=404, detail=f"Project {sutra_project} chapter {sutra_chapter} sutra {sutra_no} not found")
 
 @router.delete("/{sutra_project}/{sutra_chapter}/{sutra_no}", status_code=status.HTTP_204_NO_CONTENT)
 def delete_sutra(
@@ -90,6 +104,16 @@ def delete_sutra(
     current_admin: app_models.User = Depends(oauth2.get_current_admin),
 ):
     sutra = get_sutra_or_404(sutra_project, sutra_chapter, sutra_no, db)
-
-    db.delete(sutra)
-    db.commit()
+    if sutra:
+        db.delete(sutra)
+        db.commit()
+        return JSONResponse(status_code=status.HTTP_200_OK, content=f"Deleted sutra {sutra_project} chapter {sutra_chapter} sutra {sutra_no}")
+    else: raise HTTPException(status_code=404, detail=f"Project {sutra_project} chapter {sutra_chapter} sutra {sutra_no} not found")
+@router.delete("_by_id/{sutra_id}")
+def delete_sutra_by_id(sutra_id: int, db: Session = Depends(get_db)):
+    sutra = get_sutra_by_id_or_404(sutra_id, db)
+    if sutra:
+        db.delete(sutra)
+        db.commit()
+        return JSONResponse(status_code=status.HTTP_200_OK, content=f"Deleted sutra {sutra_id}")
+    else: raise HTTPException(status_code=404, detail=f"Project {sutra_id}not found")
